feat(db): migrate Supabase Postgres -> Turso libSQL (code conversion)
Entire DB layer moves to Turso libSQL (@libsql/client + drizzle-orm/libsql):

- db/schema.ts: pg-core -> sqlite-core. Timestamps = INTEGER ms epoch
  (timestamp_ms mode keeps the JS Date contract), jsonb -> JSON text,
  booleans -> 0/1, serial -> INTEGER PK AUTOINCREMENT, numeric -> REAL,
  text[] -> JSON text, pg enums -> typed TEXT. halfvec(3072) ->
  F32_BLOB(3072) with raw-buffer binding (embeddingToDriver/FromDriver).
- db/client.ts: @libsql/client over HTTP hrana (libsql:// -> https:),
  module-scope cache for Fluid Compute, libsqlClient() escape hatch.
- Vector search: pgvector <=>/<#> + HNSW -> vector_distance_cos() +
  DiskANN index via scripts/ops/db-create-vector-index.ts (partial,
  compress_neighbors=float8); semantic search probes vector_top_k(500)
  then filters (distance scale is now cosine, 0 = identical).
- Enrich claim: FOR UPDATE SKIP LOCKED writable CTE -> single atomic
  UPDATE ... WHERE id IN (SELECT ... LIMIT n) RETURNING.
- Dialect sweep across lib/workers/scripts/app: ILIKE -> LIKE, ::casts
  dropped/CAST(), now()/make_interval/date_trunc/to_char/extract ->
  ms arithmetic + strftime, jsonb ops -> json_extract/json_each,
  LATERAL jsonb_array_elements_text -> UNION ALL json_each branches,
  tags && ARRAY[] -> EXISTS json_each, generate_series -> JS day list,
  TRUNCATE -> DELETE, SET statement_timeout/work_mem removed,
  db().execute() -> db().all()/run() (no execute in sqlite drizzle).
- Env: POSTGRES_*/SUPABASE_* -> TURSO_DATABASE_URL/TURSO_AUTH_TOKEN
  (.env.example, lib/sources/live.ts gate, real-DB test gates).
- Tests updated in lockstep (source-inspection SQL assertions, halfvec
  test -> embedding-vector test, ms fixtures).
- Docs synced: README, docs/HANDOFF.md, architecture/ingestion.md,
  testing/strategy.md; RLS hardening doc marked superseded (no PostgREST
  surface on Turso); pg-era manual migrations removed.

scripts/migrations/pg-to-turso-copy.ts is the one-off data copier and is
removed again after cutover (needs the postgres package).

+++ scripts/ops/reset-curated-for-backfill.ts
@@ -21,7 +21,7 @@ async function main() {
   const c = db();
 
   const counts = await c
-    .select({ tier: items.tier, n: sql<number>`count(*)::int` })
+    .select({ tier: items.tier, n: sql<number>`count(*)` })
     .from(items)
     .where(
       and(inArray(items.tier, VISIBLE_ITEM_TIERS), isNotNull(items.enrichedAt)),
